pool: update relay module
Move flags, limits and stats modules to relay module
Split options module
