+++ crates/nostr-relay-pool/src/relay/flags.rs
